move more things to data/hyperdrive/ files
