+++ src/data/db_schema.py
@@ -1,4 +1,4 @@
-"""Pool info struct retured from hyperdrive contract"""
+"""Database Schemas for Basic Blockchain Datatypes.  These include things like Transactions, Accounts Etc."""
 from __future__ import annotations
 
 from decimal import Decimal
@@ -21,51 +21,7 @@ from sqlalchemy.orm import DeclarativeBase, Mapped, MappedAsDataclass, mapped_co
 class Base(MappedAsDataclass, DeclarativeBase):
     """Base class to subclass from to define the schema"""
 
-
-# TODO: Rename this to something more accurate to what is happening, e.g. HyperdriveTransactions
-class WalletInfo(Base):
-    """Table/dataclass schema for wallet information."""
-
-    __tablename__ = "walletinfo"
-
-    # Default table primary key
-    # Note that we use postgres in production and sqlite in testing, but sqlite has issues with
-    # autoincrement with BigIntegers. Hence, we use the Integer variant when using sqlite in tests
-    id: Mapped[int] = mapped_column(
-        BigInteger().with_variant(Integer, "sqlite"), primary_key=True, init=False, autoincrement=True
-    )
-
     blockNumber: Mapped[int] = mapped_column(BigInteger, ForeignKey("poolinfo.blockNumber"), index=True)
-    walletAddress: Mapped[Union[str, None]] = mapped_column(String, index=True, default=None)
-    # baseTokenType can be BASE, LONG, SHORT, LP, or WITHDRAWAL_SHARE
-    baseTokenType: Mapped[Union[str, None]] = mapped_column(String, index=True, default=None)
-    # tokenType is the baseTokenType appended with "-<maturity_time>" for LONG and SHORT
-    tokenType: Mapped[Union[str, None]] = mapped_column(String, default=None)
-    tokenValue: Mapped[Union[Decimal, None]] = mapped_column(Numeric, default=None)
-    maturityTime: Mapped[Union[int, None]] = mapped_column(BigInteger().with_variant(Integer, "sqlite"), default=None)
-    sharePrice: Mapped[Union[Decimal, None]] = mapped_column(Numeric, default=None)
-
-
-class WalletDelta(Base):
-    """Table/dataclass schema for wallet information."""
-
-    __tablename__ = "walletdelta"
-
-    # Default table primary key
-    # Note that we use postgres in production and sqlite in testing, but sqlite has issues with
-    # autoincrement with BigIntegers. Hence, we use the Integer variant when using sqlite in tests
-    id: Mapped[int] = mapped_column(
-        BigInteger().with_variant(Integer, "sqlite"), primary_key=True, init=False, autoincrement=True
-    )
-    transactionHash: Mapped[str] = mapped_column(String, ForeignKey("transactions.transactionHash"), index=True)
-    blockNumber: Mapped[int] = mapped_column(BigInteger, ForeignKey("poolinfo.blockNumber"), index=True)
-    walletAddress: Mapped[Union[str, None]] = mapped_column(String, index=True, default=None)
-    # baseTokenType can be BASE, LONG, SHORT, LP, or WITHDRAWAL_SHARE
-    baseTokenType: Mapped[Union[str, None]] = mapped_column(String, index=True, default=None)
-    # tokenType is the baseTokenType appended with "-<maturity_time>" for LONG and SHORT
-    tokenType: Mapped[Union[str, None]] = mapped_column(String, default=None)
-    delta: Mapped[Union[Decimal, None]] = mapped_column(Numeric, default=None)
-    maturityTime: Mapped[Union[Decimal, None]] = mapped_column(Numeric, default=None)
 
 
 class Transaction(Base):
@@ -85,6 +41,9 @@ class Transaction(Base):
     transactionHash: Mapped[str] = mapped_column(String, index=True, unique=True)
 
     #### Fields from base transactions ####
+    ##############
+    # NOOOOOOOOO, transaction shouldn't have a foreign key to poolInfo
+    ##############
     blockNumber: Mapped[int] = mapped_column(BigInteger, ForeignKey("poolinfo.blockNumber"), index=True)
     transactionIndex: Mapped[Union[int, None]] = mapped_column(Integer, default=None)
     nonce: Mapped[Union[int, None]] = mapped_column(Integer, default=None)
