+++ src/data/tests/test_wallet_info.py
@@ -8,7 +8,8 @@ from sqlalchemy.orm import sessionmaker
 
 import src.data.hyperdrive.postgres
 from src.data import postgres
-from src.data.db_schema import Base, WalletInfo
+from src.data.db_schema import Base
+from src.data.hyperdrive.db_schema import WalletInfo
 
 engine = create_engine("sqlite:///:memory:")  # in-memory SQLite database for testing
 Session = sessionmaker(bind=engine)
